initial push (#1)
# DX12 Base
- included visual studio dll template project
- included MIT License for the project
- included basic github readme file

# External Libraries
- Dear ImGui
- Minhook

# External Libraries
- Dear ImGui
- Minhook
Licenses can be viewed within the libs directory. Links to the respective libraries can be obtained in the readme.
Users are suggested to compile each as a "Static Library" and link to the project.
Additions to either library will be included in the main source to avoid any confusion. (i.e "ImGui::TextCentered")
*Both libraries are static compiled and will not be directly included in the project.

# Usage
include window name and compile. This was tested by injecting into an empty directX12 window.